agent/requirements.txt
python-dotenv
requests
langchain_openai
langchain
langchain_core==0.2.35
langchain_community
black
